pyproject.toml
[project]
name = "pushover"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "lawtj", email = "41483192+lawtj@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
]

[project.scripts]
pushover = "pushover:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
